feat(frontend): use libtest-mimic for planner test
