--- src/frontend/test_runner/tests/.gitignore
@@ -1,1 +1,0 @@
-gen/
