+++ src/risedevtool/Cargo.toml
@@ -21,4 +21,5 @@ serde_json = "1"
 serde_with = "1"
 serde_yaml = "0.8"
 tempfile = "3"
+workspace-hack = { version = "0.1", path = "../workspace-hack" }
 yaml-rust = "0.4"
