+++ src/storage/compactor/Cargo.toml
@@ -27,3 +27,4 @@ tokio-stream = "0.1"
 toml = "0.5"
 tonic = { version = "=0.2.0-alpha.3", package = "madsim-tonic" }
 tracing = { version = "0.1" }
+workspace-hack = { version = "0.1", path = "../../workspace-hack" }
