+++ src/storage/hummock_sdk/Cargo.toml
@@ -11,3 +11,4 @@ madsim = "=0.2.0-alpha.3"
 prost = "0.10"
 risingwave_pb = { path = "../../prost" }
 tokio = { version = "=0.2.0-alpha.3", package = "madsim-tokio", features = ["rt", "rt-multi-thread", "sync", "macros", "time", "signal"] }
+workspace-hack = { version = "0.1", path = "../../workspace-hack" }
